requirements.txt
pyTelegramBotAPI
python-dotenv
pandas
sqlalchemy
psycopg2-binary==2.9.3
#psycopg2==2.9.3
